+++ CHANGES.md
@@ -1,4 +1,3 @@
 This build requires **[IP v5.2.0](https://github.com/Efnilite/Walk-in-the-Park/releases/tag/v5.2.0)** ✨
 
-- Fixed invite not working if another plugin uses /parkour
-- Fixed error on message send
+- Fixed players being able to join lobby mode when it was not set up